+++ src/distance/distance.js
@@ -104,7 +104,7 @@ export function blindSpotTest(RC, options, toTrackDistance = false, callback) {
 
   // SPACE
   const finishFunction = async () => {
-    customButton.disabled = false
+    // customButton.disabled = false
     soundFeedback()
 
     tested += 1
@@ -162,7 +162,7 @@ export function blindSpotTest(RC, options, toTrackDistance = false, callback) {
       } else {
         // ! Reset
         tested = 0
-        customButton.disabled = true
+        // customButton.disabled = true
         // Get first response
         const firstResponse = dist[0]
         _resetCanvasLayout(
@@ -210,22 +210,22 @@ export function blindSpotTest(RC, options, toTrackDistance = false, callback) {
     }
   }
 
-  const redoFunction = () => {
-    if (!tested) return
-    tested--
-    customButton.disabled = true
+  // const redoFunction = () => {
+  //   if (!tested) return
+  //   tested--
+  //   // customButton.disabled = true
 
-    soundFeedback(3)
+  //   soundFeedback(3)
 
-    const lastResponse = dist.pop()
-    _resetCanvasLayout(
-      lastResponse.v,
-      lastResponse.closedEyeSide,
-      lastResponse.crossX,
-      true,
-      true
-    )
-  }
+  //   const lastResponse = dist.pop()
+  //   _resetCanvasLayout(
+  //     lastResponse.v,
+  //     lastResponse.closedEyeSide,
+  //     lastResponse.crossX,
+  //     true,
+  //     true
+  //   )
+  // }
 
   let arrowKeyDown = false
   let arrowIntervalFunction = null
@@ -295,7 +295,7 @@ export function blindSpotTest(RC, options, toTrackDistance = false, callback) {
 
   // Bind keys
   const bindKeysFunction = bindKeys({
-    Escape: breakFunction,
+    Escape: options.showCancelButton ? breakFunction : undefined,
     Enter: finishFunction,
     ' ': finishFunction,
     ArrowLeft: control ? arrowDownFunction : emptyFunc,
@@ -308,22 +308,23 @@ export function blindSpotTest(RC, options, toTrackDistance = false, callback) {
     },
     'keyup'
   )
-  const addedButtons = addButtons(
+
+  addButtons(
     RC.L,
     RC.background,
     {
       go: finishFunction,
-      cancel: breakFunction,
-      custom: {
-        callback: redoFunction,
-        content: phrases.RC_viewingDistanceRedo[RC.L],
-      },
+      cancel: options.showCancelButton ? breakFunction : undefined,
+      // custom: {
+      //   callback: redoFunction,
+      //   content: phrases.RC_viewingDistanceRedo[RC.L],
+      // },
     },
     RC.params.showCancelButton
   )
 
-  const customButton = addedButtons[3]
-  customButton.disabled = true
+  // const customButton = addedButtons[3]
+  // customButton.disabled = true
 
   /* -------------------------------------------------------------------------- */
   // Drag
@@ -447,7 +448,7 @@ RemoteCalibrator.prototype.measureDistance = function (options = {}, callback) {
 
   let description
   if (options.control !== undefined && options.control === false)
-    description = phrases.RC_viewingDistanceIntroNoControl[this.L]
+    description = phrases.RC_viewingDistanceIntroLiMethod[this.L]
   else description = phrases.RC_viewingDistanceIntro[this.L]
 
   options = Object.assign(
@@ -461,6 +462,7 @@ RemoteCalibrator.prototype.measureDistance = function (options = {}, callback) {
       description: description,
       check: false,
       checkCallback: false,
+      showCancelButton: true,
     },
     options
   )